add friends folder in project
