+++ py_backend/users/views.py
@@ -1,4 +1,3 @@
-from django.shortcuts import get_object_or_404, render, redirect
 from django.core.mail import send_mail
 from django.contrib.auth import authenticate, login as auth_login
 from django.contrib.auth.forms import AuthenticationForm, UserCreationForm
@@ -9,7 +8,6 @@ from django.views.decorators.http import require_http_methods
 from django.views.decorators.csrf import ensure_csrf_cookie
 from django.middleware.csrf import get_token
 from django.contrib.auth import logout
-from .models import FriendRequest
 from . import forms
 import json
 
@@ -63,22 +61,14 @@ def logout_view(request):
     logout(request)
     return JsonResponse({"status": "success"}, status=200)
 
-
-@require_http_methods(["POST"])
-def send_friend_request(request, user_id):
-    to_user = get_object_or_404(CustomUser, pk=user_id)
-    FriendRequest.objects.create(from_user=request.user, to_user=to_user, status='pending')
-    return JsonResponse({'status': 'success'}, status=200)
-
-
-@require_http_methods(["POST"])
-def tournament(request):
-    try:
-        data = json.loads(request.body.decode("utf-8"))
-    except json.JSONDecodeError:
-        return JsonResponse(data={'errors': "Invalid JSON format"}, status=406)
+# @require_http_methods(["POST"])
+# def tournament(request):
+#     try:
+#         data = json.loads(request.body.decode("utf-8"))
+#     except json.JSONDecodeError:
+#         return JsonResponse(data={'errors': "Invalid JSON format"}, status=406)
     
-    return JsonResponse({"status": "success"}, status=200)
+#     return JsonResponse({"status": "success"}, status=200)
 
 
 @require_http_methods(["GET"])
